split to other tests spec
